feat(Sede): elimina sede_info para  incluirla en el mismo fichero ya que ésta no se usa externamente - Resuelve #10
